Fix a bug in hash_fnv1.c and hash_fnv1a.c
The hash_fnv1.c and hash_fnv1a.c routines contained a bug where the
routine would hash the initial octet of the key over and over again.  They
also failed to restrict the output to 32-bits.  This commit fixes both
problems and adds tests to t_hashtab.c to validate that the correct values
are computed.