Build system cleanups
This commit cleans up a couple minor things from the build system.  First,
the AC_DEFINE_VERSION macro is altered to use the $PACKAGE_VERSION shell
variable (it is expected a future refinement will alter this to use an
explicit version number).  Second, left-over references to dbprim-config
are removed from the Makefile.am.