Add missing shortcircuit test case
This change adds a missing shortcircuit test case that I never noticed
was missing.  I also tracked down why my local environment was not seeing
these: CFLAGS was getting set to a variant with "-O2", so the lines were
being optimized out.  To ensure this never happens to the travis testing,
a "CFLAGS=" value is added to the ./configure invocation.